api/generate-pptx/requirements.txt
python-pptx>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
